Fri, Sep 6, 2024, 9:27 AM +05:30

+++ CS50p/pset3/fuel/fuel.py
@@ -28,7 +28,7 @@ def get_fraction(string):
 
         try:
             numer, denom = fraction.split("/")
-            numer, denom = float(numer), float(denom)
+            numer, denom = int(numer), int(denom)
             return (numer/denom)
         except ValueError:
             print("Input format 'X/Y' and X,Y are numbers")
